#1125: more schema field cleanup
