--- src/components/fields/BlockField.scss
@@ -1,30 +1,0 @@
-/*!
- * Copyright (C) 2021 PixieBrix, Inc.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-.BlockCard {
-  &.invalid {
-    border: 1px solid #fe7c96;
-  }
-
-  &.card {
-    border-radius: 0;
-
-    .card-body {
-      padding: 1rem;
-    }
-  }
-}
