+++ src/components/fields/schemaFields/SelectorSelectorField.tsx
@@ -15,41 +15,24 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import React, { useMemo } from "react";
-import Select from "react-select";
-import { IBlock } from "@/core";
+import React from "react";
+import { SchemaFieldProps } from "@/components/fields/schemaFields/propTypes";
+import { fieldLabel } from "@/components/fields/fieldUtils";
+import ConnectedFieldTemplate from "@/components/form/ConnectedFieldTemplate";
+import SelectorSelectorWidget from "@/devTools/editor/fields/SelectorSelectorWidget";
 
-interface BlockOption {
-  value: string;
-  label: string;
-  block: IBlock;
-}
-
-const BlockSelector: React.FunctionComponent<{
-  onSelect: (service: IBlock) => void;
-  blocks: IBlock[];
-  placeholder?: string;
-}> = ({ onSelect, blocks, ...props }) => {
-  const blockOptions = useMemo(
-    () =>
-      (blocks ?? []).map((x) => ({
-        value: x.id,
-        label: x.name,
-        block: x,
-      })),
-    [blocks]
-  );
-
+const SelectorSelectorField: React.FunctionComponent<
+  SchemaFieldProps<string>
+> = (props) => {
+  const { label, name, schema } = props;
   return (
-    <Select
-      key={Math.random()}
-      options={blockOptions}
-      onChange={(x: BlockOption) => {
-        onSelect(x.block);
-      }}
+    <ConnectedFieldTemplate
+      label={label ?? fieldLabel(name)}
+      description={schema.description}
+      as={SelectorSelectorWidget}
       {...props}
     />
   );
 };
 
-export default BlockSelector;
+export default SelectorSelectorField;
