+++ src/components/fields/schemaFields/widgets/TextWidget.tsx
@@ -16,7 +16,7 @@
  */
 
 import React, { useMemo, useState } from "react";
-import { SchemaFieldProps } from "@/components/fields/propTypes";
+import { SchemaFieldProps } from "@/components/fields/schemaFields/propTypes";
 import { useField } from "formik";
 import Select, { OptionsType } from "react-select";
 import { sortBy, uniq } from "lodash";
