+++ src/devTools/editor/fields/FormRendererOptions.tsx
@@ -15,17 +15,13 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { FieldRenderer } from "@/components/fields/blockOptions";
+import SchemaField from "@/components/fields/schemaFields/SchemaField";
 import FormBuilder from "@/components/formBuilder/FormBuilder";
 import { Schema } from "@/core";
 import React from "react";
+import { OUTPUT_KEY_SCHEMA } from "@/components/fields/schemaFields/genericOptionsFactory";
 
 export const FORM_RENDERER_ID = "@pixiebrix/form";
-
-const outputKeySchema: Schema = {
-  type: "string",
-  description: "A name to refer to this brick in subsequent bricks",
-};
 
 const recordIdSchema: Schema = {
   type: "string",
@@ -43,13 +39,13 @@ const FormRendererOptions: React.FC<{
     <div>
       <FormBuilder name={configName} />
 
-      <FieldRenderer name={`${configName}.recordId`} schema={recordIdSchema} />
+      <SchemaField name={`${configName}.recordId`} schema={recordIdSchema} />
 
       {showOutputKey && (
-        <FieldRenderer
+        <SchemaField
           name={`${name}.outputKey`}
           label="Output Variable"
-          schema={outputKeySchema}
+          schema={OUTPUT_KEY_SCHEMA}
         />
       )}
     </div>
