+++ src/devTools/editor/fields/SelectorSelectorField.tsx
@@ -15,234 +15,29 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import React, { useCallback, useContext, useMemo, useState } from "react";
+import React, { useMemo } from "react";
 import { useField } from "formik";
-import { compact, isEmpty, sortBy, uniqBy } from "lodash";
-import { Button } from "react-bootstrap";
-import { faMousePointer } from "@fortawesome/free-solid-svg-icons";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import {
-  disableOverlay,
-  enableSelectorOverlay,
-  selectElement,
-} from "@/background/devtools";
-import { DevToolsContext } from "@/devTools/context";
-import { SelectMode } from "@/nativeEditor/selector";
-import { ElementInfo } from "@/nativeEditor/frameworks";
-import { Framework } from "@/messaging/constants";
-import CreatableAutosuggest, {
-  SuggestionTypeBase,
-} from "@/devTools/editor/fields/creatableAutosuggest/CreatableAutosuggest";
-import SelectorListItem from "@/devTools/editor/fields/selectorListItem/SelectorListItem";
-import { getErrorMessage } from "@/errors";
-import useNotifications from "@/hooks/useNotifications";
 import ComplexObjectValue from "@/components/fields/schemaFields/widgets/ComplexObjectWidget";
-
-interface ElementSuggestion extends SuggestionTypeBase {
-  value: string;
-  elementInfo?: ElementInfo;
-}
-
-function getSuggestionsForElement(
-  elementInfo: ElementInfo | undefined
-): ElementSuggestion[] {
-  if (!elementInfo) {
-    return [];
-  }
-
-  return uniqBy(
-    compact([
-      ...(elementInfo.selectors ?? []).map((value) => ({ value, elementInfo })),
-      ...getSuggestionsForElement(elementInfo.parent),
-    ]).filter(
-      (suggestion) => suggestion.value && suggestion.value.trim() !== ""
-    ),
-    (suggestion) => suggestion.value
-  );
-}
-
-function renderSuggestion(suggestion: ElementSuggestion): React.ReactNode {
-  return (
-    <SelectorListItem
-      value={suggestion.value}
-      hasData={suggestion.elementInfo.hasData}
-      tag={suggestion.elementInfo.tagName}
-    />
-  );
-}
-
-interface CommonProps {
-  initialElement?: ElementInfo;
-  framework?: Framework;
-  selectMode?: SelectMode;
-  traverseUp?: number;
-  isClearable?: boolean;
-  sort?: boolean;
-  root?: string;
-  disabled?: boolean;
-}
-
-export const SelectorSelectorControl: React.FunctionComponent<
-  CommonProps & {
-    value: string;
-    onSelect: (selector: string) => void;
-  }
-> = ({
-  value,
-  onSelect,
-  initialElement,
-  framework,
-  selectMode = "element",
-  traverseUp = 0,
-  isClearable = false,
-  sort = false,
-  root,
-  disabled = false,
-}) => {
-  const { port } = useContext(DevToolsContext);
-  const notify = useNotifications();
-  const [element, setElement] = useState(initialElement);
-  const [isSelecting, setSelecting] = useState(false);
-
-  const suggestions: ElementSuggestion[] = useMemo(() => {
-    const raw = getSuggestionsForElement(element);
-    return sort ? sortBy(raw, (x) => x.value.length) : raw;
-  }, [element, sort]);
-
-  const enableSelector = useCallback(
-    (selector: string) => {
-      try {
-        void enableSelectorOverlay(port, selector);
-      } catch {
-        // The enableSelector function throws errors on invalid selector
-        // values, so we're eating everything here since this fires any
-        // time the user types in the input.
-      }
-    },
-    [port]
-  );
-
-  const disableSelector = useCallback(() => {
-    void disableOverlay(port);
-  }, [port]);
-
-  const onHighlighted = useCallback(
-    (suggestion: ElementSuggestion | null) => {
-      if (suggestion) {
-        enableSelector(suggestion.value);
-      } else {
-        disableSelector();
-      }
-    },
-    [enableSelector, disableSelector]
-  );
-
-  const onTextChanged = useCallback(
-    (value: string) => {
-      disableSelector();
-      enableSelector(value);
-      onSelect(value);
-    },
-    [disableSelector, enableSelector, onSelect]
-  );
-
-  const select = useCallback(async () => {
-    setSelecting(true);
-    try {
-      const selected = await selectElement(port, {
-        framework,
-        mode: selectMode,
-        traverseUp,
-        root,
-      });
-
-      if (isEmpty(selected)) {
-        notify.error("Unknown error selecting element", {
-          error: new Error("selectElement returned empty object"),
-        });
-        return;
-      }
-
-      setElement(selected);
-
-      const selectors = selected.selectors ?? [];
-
-      const firstSelector = (sort
-        ? sortBy(selectors, (x) => x.length)
-        : selectors)[0];
-
-      console.debug("Setting selector", { selected, firstSelector });
-      onSelect(firstSelector);
-    } catch (error: unknown) {
-      notify.error(`Error selecting element: ${getErrorMessage(error)}`, {
-        error,
-      });
-    } finally {
-      setSelecting(false);
-    }
-  }, [
-    port,
-    sort,
-    framework,
-    notify,
-    setSelecting,
-    traverseUp,
-    selectMode,
-    setElement,
-    onSelect,
-    root,
-  ]);
-
-  return (
-    <div className="d-flex">
-      <div>
-        <Button
-          onClick={select}
-          disabled={isSelecting || disabled}
-          variant="info"
-          aria-label="Select element"
-        >
-          <FontAwesomeIcon icon={faMousePointer} />
-        </Button>
-      </div>
-      <div className="flex-grow-1">
-        <CreatableAutosuggest
-          isClearable={isClearable}
-          isDisabled={isSelecting || disabled}
-          suggestions={suggestions}
-          inputValue={value}
-          inputPlaceholder="Choose a selector..."
-          renderSuggestion={renderSuggestion}
-          onSuggestionHighlighted={onHighlighted}
-          onSuggestionsClosed={disableSelector}
-          onTextChanged={onTextChanged}
-        />
-      </div>
-    </div>
-  );
-};
+import SelectorSelectorWidget, {
+  SelectorSelectorProps,
+} from "@/devTools/editor/fields/SelectorSelectorWidget";
+import ConnectedFieldTemplate from "@/components/form/ConnectedFieldTemplate";
 
 const SelectorSelectorField: React.FunctionComponent<
-  CommonProps & { name?: string }
-> = ({ name, ...props }) => {
+  SelectorSelectorProps & { name?: string }
+> = (props) => {
   // Some properties (e.g., the menuItem's container prop) support providing an array of selectors.
   // See awaitElementOnce for for the difference in the semantics vs. nested CSS selectors
-  const [field, , helpers] = useField<string | string[]>(name);
+  const [field] = useField<string | string[]>(props.name);
 
-  if (Array.isArray(field.value)) {
-    return (
-      // Match what we show in other places interface. Not a good UX, but it's consistent!
-      <ComplexObjectValue name={name} schema={null} />
-    );
-  }
+  const isArray = Array.isArray(field.value);
 
-  return (
-    <SelectorSelectorControl
-      value={field.value}
-      onSelect={helpers.setValue}
-      {...props}
-    />
+  const Widget = useMemo(
+    () => (isArray ? ComplexObjectValue : SelectorSelectorWidget),
+    [isArray]
   );
+
+  return <ConnectedFieldTemplate {...props} as={Widget} />;
 };
 
 export default SelectorSelectorField;
